backend/requirements/requirements-dev.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements/requirements-dev.in
alembic==1.15.1
    # via -r requirements/requirements-dev.in
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
argon2-cffi==23.1.0
    # via -r requirements/requirements-dev.in
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
authlib==1.5.1
    # via -r requirements/requirements-dev.in
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
cffi==1.17.1
    # via
    #   argon2-cffi-bindings
    #   cryptography
click==8.1.8
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
cryptography==44.0.2
    # via authlib
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
fastapi==0.115.11
    # via -r requirements/requirements-dev.in
fastapi-cli==0.0.7
    # via fastapi
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via fastapi
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
iniconfig==2.0.0
    # via pytest
jinja2==3.1.6
    # via fastapi
loguru==0.7.3
    # via -r requirements/requirements-dev.in
mako==1.3.9
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
packaging==24.2
    # via pytest
pluggy==1.5.0
    # via pytest
pycparser==2.22
    # via cffi
pydantic==2.10.6
    # via fastapi
pydantic-core==2.27.2
    # via pydantic
pygments==2.19.1
    # via rich
pytest==8.3.5
    # via
    #   -r requirements/requirements-dev.in
    #   pytest-asyncio
pytest-asyncio==0.25.3
    # via -r requirements/requirements-dev.in
python-environ==0.4.54
    # via -r requirements/requirements-dev.in
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.2
    # via uvicorn
rich==13.9.4
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.13.2
    # via fastapi-cli
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.39
    # via
    #   -r requirements/requirements-dev.in
    #   alembic
starlette==0.46.1
    # via fastapi
typer==0.15.2
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   sqlalchemy
    #   typer
uvicorn==0.34.0
    # via
    #   fastapi
    #   fastapi-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.4
    # via uvicorn
websockets==15.0.1
    # via uvicorn
